requirements.txt
django
django-crispy-forms
django-debug-toolbar
pillow
psycopg2
pytz
requests
